+++ app/src/Controller/Oauth2.php
@@ -2,7 +2,7 @@
 
 namespace Dappur\Controller;
 
-use Interop\Container\ContainerInterface;
+use Psr\Container\ContainerInterface as ContainerInterface;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;